+++ R/f_simulation_base_survival.R
@@ -509,7 +509,7 @@ getSimulationSurvival <- function(
         eventTime = eventTime,
         kappa = kappa,
         delayedResponseAllowed = TRUE,
-        .pi1Default = C_PI_1_DEFAULT,
+        .pi1Default = .getPi1Default(type = "power", endpoint = "survival"),
         .silent = TRUE
     )